Install solr extension from source
Using 2.4.0 release + custom patch (from macports project) seems to
be the only working combination right now. Read the comments in the code
and the links there.

Resolves solr issue in #16, waiting for the upstream as described in #19

This is basically a backport of #58 that was applied
to php72 and up.

The problem with php71 was detected @ https://github.com/moodlehq/moodle-docker/pull/134#issuecomment-691329406