Same 8.0.0RC5 but moved back to pecl instead of pickle
It seems that, at the end, pear/pecl continue being available
for php80, no matter it was deprecated for php74. So, as far
as pickle was unofficial/worse alternative, moving back to pecl.

I've left some notes about pickle in the Dockerfile so we can
easily move to it again in the future if it becomes better and
supported tool.

Bumping to 8.0.0RC5

- oci8 support added.
- mongodb support added.
- sqlsrv support added (5.9.0preview1)

Still some extensions not available:
- xmlrpc : -- not existing as of 8.0.0RC5 (not yet @ PECL, https://php.watch/versions/8.0#xmlrpc)