First cut to moodle-php-apache multiarch builder
First attempt to switch the automated building of moodle-php-apache
images from DockerHub to GHA. Main reason being that the former doesn't
support multiarch builds (only multiarch storage).

Link: https://github.com/docker/hub-feedback/issues/1874

Experimentally we'll be, also, sending the images to both DockerHub
and GitHub registries in case we need to switch some day.

This requires:

1. Some secrets to be created:
  - DOCKERHUB_USERNAME
  - DOCKERHUB_TOKEN (write perms)
  - DOCKERHUB_PASSWORD, grrr, descriptions cannot be updated using PATs,
    see https://github.com/peter-evans/dockerhub-description/issues/10
  - GH_USERNAME
2. Autobuilds @ DockerHub to be disabled (this takes exactly on them)

Remove old dockerhub hook, because now GHA is in charge of tagging.