Switch to v3 that supports editing descriptions with token and 2fa
Previously, the dockerhub description only was editable with
user and password, but it was not working for accounts having
2FA enabled (like ours). Hence the job in master was always
failing with 401.

In July 2022 Github made the needed change to update dockerhub
descriptions using PAT, that is independent of the 2FA sttings.

Hence, switching to v3 of the action and moving to token.