New workflow to trigger automatic image builds
This is executed daily and performs various checks:

- Is there a newer timezonedb extension available.
- (more coming soon).

And, if there is any of the checks happening, it automatically
dispatches a reposity_dispatch (auto-build-triggered) event,
causing the test and build workflow to be triggered.

Now the test & build workflow supports workflow_call events

That way we can make other workflows in the repository to,
conditionally, launch a rebuild when some conditions are met:
- When there is a new version of an extension.
- When there is a new PHP release.
- After X days
- ...

Note this is related to https://tracker.moodle.org/browse/MDL-76675

Also, update some actions to use new nodejs versions and avoid
some deprecation warnings.