Rename moodle ini to support easier lexical sorring
With the new entrypoint conf.d support, we want to be able to easily
support prioritisation of the various ini options.

Numerically prefixed ini files are easier to lexically order.

Typically we want the standard file to be at the beginning because it's
a staple which others should be able to override in most cases, but we
also want to allow people to apply earlier files too.

As it stands, the current name (docker-php-moodle.ini) will be called
after any numerically-prefixed ini file due to the standard lexical
sort.