Make the 8.2.3 images sticky
It has been detected that with the latest 8.2.4 images built
2 days ago, there is some change in the apache configuration
causing behat to be unable to perform reruns correctly.

The message when running behat is:

is not a behat test site

And, in the apache logs there are hundreds of:

Symbolic link not allowed

It has been checked that the problem is something recent, sure
something changed in apache conf. And that 8.2.3 images are
working ok. Hence switching to them while we investigate the
differences.