Trigger an image rebuild if there are new php releases
Similarly to what we already do, rebuilding images when
a new timezonedb version is detected, this commit adds
a new check looking for the existence of new PHP releases
using this URL: https://www.php.net/releases/?json&version=8.0-anything

Note that it may be the case of the new version already released
and the docker image still not being available, but that doesn't
matter much, because in the next execution (currently every 3 days)
it will be tried and rebuilt again.

Also, worth commenting that we have had to filter the "docker run"
commands with a "tail -1" because there is other output related
with configuration and entry points that is breaking the expected
output.